rename filenames and refactor codes to make them more pythonic
