+++ models/opnilog/inference.py
@@ -2,7 +2,7 @@
 import logging
 
 # Third Party
-from OpniLogParser import LogParser
+from opnilog_parser import LogParser
 
 
 def init_model(save_path="output/"):
